modal and search functionality added

+++ src/components/dashboard/SingleProductCardDashboard.jsx
@@ -1,32 +1,41 @@
 /* eslint-disable react/prop-types */
-import toast from "react-hot-toast";
-import { Link } from "react-router-dom";
-import "./../../components/home/products.css";
+import toast from 'react-hot-toast';
+import { Link } from 'react-router-dom';
+import './../../components/home/products.css';
+import { useState } from 'react';
+import ConfirmDeleteModal from '../ConfirmDeleteModal';
 
-// eslint-disable-next-line react/prop-types
 const SingleProductCardDashboard = ({ product, onDelete }) => {
-  const token = localStorage.getItem("token");
+  const token = localStorage.getItem('token');
   const { _id, title, brand, price, description, image_url } = product;
+  const [showModal, setShowModal] = useState(false);
 
   const handleDelete = async () => {
     await fetch(`https://cric-expo-server1.onrender.com/products/${_id}`, {
-      method: "DELETE",
+      method: 'DELETE',
       headers: {
-        "Content-type": "application/json",
+        'Content-type': 'application/json',
         authorization: `Bearer ${token}`,
       },
     })
       .then((res) => res.json())
       .then(() => {
-        toast.success("Product Deleted");
+        toast.success('Product Deleted');
         onDelete(_id);
       });
+  };
+
+  const handleCloseModal = () => setShowModal(false);
+  const handleShowModal = () => setShowModal(true);
+  const handleConfirmDelete = () => {
+    handleDelete();
+    handleCloseModal();
   };
 
   return (
     <div className="card w-96 bg-base-100 shadow-xl all_product_section">
       <figure>
-        <img src={image_url} alt="Shoes" />
+        <img src={image_url} alt={title} />
       </figure>
       <div className="card-body">
         <h2 className="card-title">{title}</h2>
@@ -40,11 +49,20 @@ const SingleProductCardDashboard = ({ product, onDelete }) => {
           <button className="btn bg-yellow-500 text-black">
             <Link to={`edit/${_id}`}>Edit</Link>
           </button>
-          <button onClick={handleDelete} className="btn bg-red-500 text-black">
+          <button
+            onClick={handleShowModal}
+            className="btn bg-red-500 text-black"
+          >
             Delete
           </button>
         </div>
       </div>
+
+      <ConfirmDeleteModal
+        show={showModal}
+        handleClose={handleCloseModal}
+        handleConfirm={handleConfirmDelete}
+      />
     </div>
   );
 };
